requirements.txt
CairoSVG~=2.7.1
pillow~=10.3.0
svgpathtools~=1.6.1
uuid~=1.30
